# lib/api-zod/src/generated/types/edaStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EdaStatsClassDistribution } from "./edaStatsClassDistribution";
import type { EdaStatsCorrelationMatrix } from "./edaStatsCorrelationMatrix";
import type { EdaStatsDatasetInfo } from "./edaStatsDatasetInfo";
import type { EdaStatsFeatureImportances } from "./edaStatsFeatureImportances";
import type { EdaStatsHourlyDistributionItem } from "./edaStatsHourlyDistributionItem";
import type { EdaStatsModelMetrics } from "./edaStatsModelMetrics";
import type { EdaStatsMonthlyStats } from "./edaStatsMonthlyStats";
import type { EdaStatsScatterDataItem } from "./edaStatsScatterDataItem";

export interface EdaStats {
  dataset_info: EdaStatsDatasetInfo;
  class_distribution: EdaStatsClassDistribution;
  monthly_stats: EdaStatsMonthlyStats;
  correlation_matrix: EdaStatsCorrelationMatrix;
  feature_importances: EdaStatsFeatureImportances;
  model_metrics: EdaStatsModelMetrics;
  scatter_data: EdaStatsScatterDataItem[];
  hourly_distribution: EdaStatsHourlyDistributionItem[];
}