# lib/api-zod/src/generated/types/predictRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PredictRequest {
  /** Temperature in Celsius */
  temp_c: number;
  /** Dew Point Temperature in Celsius */
  dew_point_temp_c: number;
  /** Relative Humidity percentage */
  rel_hum: number;
  /** Wind Speed in km/h */
  wind_speed_kmh: number;
  /** Visibility in km */
  visibility_km: number;
  /** Pressure in kPa */
  press_kpa: number;
}