# lib/api-zod/src/generated/types/predictResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PredictResponseProbabilities } from "./predictResponseProbabilities";

export interface PredictResponse {
  predicted_weather: string;
  probabilities: PredictResponseProbabilities;
  confidence: number;
}